requirements.txt
git+https://github.com/openai/whisper.git 
gradio
ffmpeg
faster-whisper
ctranslate2 OpenNMT-py sentencepiece
